requirements.txt
altair==5.5.0
annotated-types==0.7.0
anyio==4.10.0
asttokens==2.4.1
attrs==25.3.0
backcall==0.2.0
blinker==1.9.0
cachetools==6.1.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
colorama==0.4.6
comm==0.2.1
debugpy==1.8.0
decorator==5.1.1
distro==1.9.0
entrypoints==0.4
exceptiongroup @ file:///home/conda/feedstock_root/build_artifacts/exceptiongroup_1746947292760/work
executing==2.0.1
fastapi==0.116.1
filelock==3.19.1
Flask==3.1.1
fsspec==2025.7.0
gitdb==4.0.12
GitPython==3.1.45
groq==0.31.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.34.4
idna==3.10
importlib_metadata @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_importlib-metadata_1747934053/work
ipykernel @ file:///D:/bld/ipykernel_1754352877088/work
ipython==8.21.0
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.6
jiter==0.10.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
jupyter_client==8.6.0
jupyter_core==5.7.1
langchain-core==0.3.74
langchain-groq==0.3.7
langchain-openai==0.3.30
langserve==0.3.1
langsmith==0.4.14
MarkupSafe==3.0.2
matplotlib-inline==0.1.6
narwhals==2.1.2
nest-asyncio==1.6.0
numpy==2.2.6
openai==1.99.9
orjson==3.11.2
packaging @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_packaging_1745345660/work
pandas==2.3.1
parso==0.8.3
pickleshare==0.7.5
pillow==11.3.0
platformdirs==4.2.0
prompt-toolkit==3.0.43
protobuf==6.32.0
psutil==5.9.8
pure-eval==0.2.2
pyarrow==21.0.0
pydantic==2.11.7
pydantic_core==2.33.2
pydeck==0.9.1
Pygments==2.17.2
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==1.1.1
pytz==2025.2
pywin32 @ file:///D:/bld/bld/rattler-build_pywin32_1752564081/work
PyYAML==6.0.2
pyzmq==25.1.2
referencing==0.36.2
regex==2025.7.34
requests==2.32.4
requests-toolbelt==1.0.0
rpds-py==0.27.0
safetensors==0.6.2
six==1.16.0
smmap==5.0.2
sniffio==1.3.1
sse-starlette==3.0.2
stack-data==0.6.3
starlette==0.47.2
streamlit==1.48.1
tenacity==9.1.2
tiktoken==0.11.0
tokenizers==0.21.4
toml==0.10.2
tornado==6.4
tqdm==4.67.1
traitlets==5.14.1
transformers==4.55.2
typing-inspection==0.4.1
typing_extensions @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_typing_extensions_1751643513/work
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.35.0
watchdog==6.0.0
wcwidth==0.2.13
Werkzeug==3.1.3
zipp @ file:///home/conda/feedstock_root/build_artifacts/zipp_1749421620841/work
zstandard==0.23.0
